# api/docs/swagger.yaml
definitions:
...
  models.CreateDoctorType:
    properties:
      clinic_branch_id:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
  models.CreateJournal:
    properties:
      article:
        type: string
      author_id:
        type: string
      theme:
        type: string
    type: object
...
  models.DoctorType:
    properties:
      clinic_branch_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.DoctorTypesResponse:
    properties:
      count:
        type: integer
      doctor_types:
        items:
          $ref: '#/definitions/models.DoctorType'
        type: array
    type: object
...
  models.Journal:
    properties:
      article:
        type: string
      author_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      theme:
        type: string
      updated_at:
        type: string
    type: object
  models.JournalsResponse:
    properties:
      count:
        type: integer
      journals:
        items:
          $ref: '#/definitions/models.Journal'
        type: array
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
...
  models.UpdateDoctorType:
    properties:
      clinic_branch_id:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
  models.UpdateJournal:
    properties:
      article:
        type: string
      author_id:
        type: string
      theme:
        type: string
    type: object
...
info:
  contact: {}
  description: Online doctor appointments and drug orders
  title: ShifoLink
  version: 1.0.0
paths:
...
  /doctor_type:
    post:
      consumes:
      - application/json
      description: Create a new doctor type
      parameters:
      - description: doctor type data
        in: body
        name: doctor_type
        required: true
        schema:
          $ref: '#/definitions/models.CreateDoctorType'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DoctorType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new doctor type
      tags:
      - doctor_type
  /doctor_type/{id}:
    delete:
      consumes:
      - application/json
      description: Delete doctor type
      parameters:
      - description: doctor_type id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete doctor type
      tags:
      - doctor_type
    get:
      consumes:
      - application/json
      description: Get doctor type by id
      parameters:
      - description: doctor_type
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DoctorType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get doctor type by id
      tags:
      - doctor_type
    put:
      consumes:
      - application/json
      description: Update doctor type by id
      parameters:
      - description: doctor type id
        in: path
        name: id
        required: true
        type: string
      - description: doctor_type
        in: body
        name: doctor_type
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDoctorType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DoctorType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update doctor type by id
      tags:
      - doctor_type
  /doctor_types:
    get:
      consumes:
      - application/json
      description: Get doctor_types list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DoctorTypesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get doctor_types list
      tags:
      - doctor_type
...
  /journal:
    post:
      consumes:
      - application/json
      description: Create a new journal
      parameters:
      - description: journal data
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/models.CreateJournal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Journal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new journal
      tags:
      - journal
  /journal/{id}:
    delete:
      consumes:
      - application/json
      description: Delete journal
      parameters:
      - description: journal id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete journal
      tags:
      - journal
    get:
      consumes:
      - application/json
      description: Get journal by id
      parameters:
      - description: journal
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Journal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get journal by id
      tags:
      - journal
    put:
      consumes:
      - application/json
      description: Update journal by id
      parameters:
      - description: journal id
        in: path
        name: id
        required: true
        type: string
      - description: journal
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/models.UpdateJournal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Journal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update journal by id
      tags:
      - journal
  /journals:
    get:
      consumes:
      - application/json
      description: Get journals list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JournalsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get journals list
      tags:
      - journal
...
swagger: "2.0"